Abort when using GCL from replicated_collectives
Summary:
We will lose the ability to call GCL soon, when GCL_LIB_PATH is
deprecated. The replicated_collectives tool has been ported to GCL in
D31360. In the future we will remove it from poplibs altogether, but
before that we need to make sure all the poplibs tests are run by GCL.

Reviewers: richard, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, larsph

Reviewed By: richard, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T26156

Differential Revision: https://phabricator.sourcevertex.net/D31361