Improve collective dynamic slice efficiency
Summary:
With this change all dynamic slice operations use 1D dynamic slice
vertices which are more efficient. Previously we used 2D dynamic slice
vertices since the fragments were reordered in-between creating the
sliceable tensor and slicing into it.

This is fixed by pushing the reordering of fragments by ring index up the
call stack to the point where the tensor is created and by changing the
initial ring index used for reduce-scatter so there is no need to apply
an offset when the fragments are accessed.

Depends on D31487

Fixes T26216

Reviewers: saml, samuelh, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: saml, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T26216

Differential Revision: https://phabricator.sourcevertex.net/D31488