Use the best available supported codelet given block size
Summary:
Convert CSR, CSC and COO formats with larger
block sizes to smaller ones. This is the basis on which
support for larger block sizes for which efficient codelets
are not supported can be provided.

Implementation meta information buckets are then
converted to the original block size so the block size
conversion on the device is transparent to the host.

Fixes: T24987

Reviewers: tomm, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk

Maniphest Tasks: T24987

Differential Revision: https://phabricator.sourcevertex.net/D31367