Adding configurable cell order to LSTM interface
Summary:
This change adds an external shape to the weights and tensors for an lstm.

A new option in the LstmParams allows the user to specify the order of the outermost dimension which is the different gates inside of the op.

The allocated shape is unchanged, the only difference is the returned shape.

Test Plan: new tests using the ordering used by pytorch

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, alfiee, godfrey.da.costa, timp, tomm, oleksiik, jorgec, samk

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oleksiik, jorgec

Subscribers: georgep, jorgec

Maniphest Tasks: T25539

Differential Revision: https://phabricator.sourcevertex.net/D31328