Adding configurable cell order to GRU interface
Summary:
GRU counterpart to [[ https://phabricator.sourcevertex.net/D31328 | D31328 ]]

This change adds an external shape to the weights and tensors for a gru.

A new option in the GruParams allows the user to specify the order of the outermost dimension which is the different gates inside of the op.

The allocated shape is unchanged, the only difference is the returned shape.

Test Plan:
new tests using a non-default ordering

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk, tomm, jorgec, oleksiik

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oleksiik

Maniphest Tasks: T25539

Differential Revision: https://phabricator.sourcevertex.net/D31821