Codelets for Sparse to dense multi slice and dense to sparse update (C++ only)
Summary:
Functioning but some points TODO noted in the code - they can be revised as the assembly is developed.

- Requires C++ 17 to allow use of constexpr in the vertex code - needed due to types on vertex being InOut Input etc.  This change is in a poplar diff but not landed yet.

Changes:
Moved: Code from SparseDenseMatMulElementWise.cpp so that sparse metadata and NZ data generation can be shared.

Added: A slice codelet that is given buckets of sparse data, and a tensor of rows to retrieve from the sparse data. It is given a subGroupID and a rowOffset which tells it to process only the data in the buckets with a matching subGroupID, and when that is found, interpret it given the rowOffset.
The vertex searches through the buckets and subGroups in each bucket to find data with matching subGroupID - then extracting the required rows found in the sparse data and populating the subT result tensor. Populating is an update of the new values found (As the vertex may be called multiple times and will populate part of the sparse data each call), or may find multiple buckets containing data with the same subGroupID .

Added: Update Add codelet that does the reverse.  When indexed row data is found in the dense subT tensor it is copied into the NZ data  in baseT using result = result +scale*input.  Only the subT values that are represented in the sparse baseT are updated.

Ref T20690 (Assembler to be added)

Test Plan: Added codelet specific tests covering all 4 codelet types (Slice, UpdateAdd with float, half variants.  Simple until assembler is created

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Maniphest Tasks: T20690

Differential Revision: https://phabricator.sourcevertex.net/D31616