C++/ASM codelets to do block transposes suitable for direct loading of coefficients in GradW AMP codelet
Summary:
The existing GradW AMP codelet

The existing supervisor codelet for transposition divides number of
transpositions between workers. This results in poor thread balance.

Transpose the input such that the output is laid out for the GradW
AMP codelet to read linearly. The outer dimension of the input must
be a multiple of 16 for halves and 8 for floats.

Fixes T26820 T26823 T26821

Test Plan: New set of tests added

Reviewers: tomm, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: tomm, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T26823, T26821, T26820

Differential Revision: https://phabricator.sourcevertex.net/D32436