Use of "Multiply and shift" to divide in the sparse-dense multislice codelets.  Use and renaming of offsetToFirstOutputEntry
Summary:
As per the title.  The method of achieving divide is to:
Determine a constant at compile time `reciprocalMulFactor= 1+(1<<reciprocalMulShift)/divisor`, passed in vertex state.  Then at run time we can divide by a constant using `out = (in*reciprocalMulFactor) >>reciprocalMulShift`
This is a standard technique, frequently used in poplibs to divide work - and so dividing by 3, 6, 12 or 24.
The technique does have limitations, and for greater flexibility the shift can be chosen to suit the divisor, however in this case `reciprocalMulShift = 16` is suitable and simplifies matters.

In this case we know that `in` is within a 16 bit unsigned range, and is generated by multiplication of 2 integers (divisor and the input that we want to extract) and is so divisible exactly.  It is possible to check the calculation over the full range of values that fit that criteria and compare to a division on the host.  Additionally the intermediate value `in*reciprocalMulFactor` will be within a 32 bit range.

Ref T20690

Test Plan: Further tests of the codelet using various z-scaling parameters to verify this functions correctly.

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm, oleksiik

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Maniphest Tasks: T20690

Differential Revision: https://phabricator.sourcevertex.net/D32819