Embedding layer with sparse to dense slice and dense to sparse update
Summary:
Slice and update add functions for an embedding layer, where the vocabulary is a sparse tensor generated by a matrix multiply.  The extracted slices are expanded into dense tensors and in update the input tensor with which to update the sparse vocabulary is also a dense tensor.

The splits in X,Y,Z are extracted from the matmul plan, as is the tile and subGroupID sequence - using the same logic/functions as the matmul plan.  The memory layout of the sparse data is left un-touched, and the arrangement of slices (the output of the embbedding slice) is arranged consistently with that plan.  This means that we maintain X-splits (columns are split in the same way) and partition the slices over the remaining tile groups.

MAJOR LIMITATIONS:
Assumption that bucket content is not spilled to neighbouring buckets in the X,Y dimensions.(That is - there is no support for bucket rotation)
Update gathers all slices and so will have a large temporary memory requirement.
Even in slice there is a reliance on the matmul picking a helpful number of X-splits so that temporary memory is acceptable.  The dense embedding layer chooses this partition for itself.  Forcing X-splits in the matmul plan, or adding an option to tell it about the requirement should be enough to solve this issue for slice, and could help update.  However for update further improvements are necessary as the temporary memory problem is bigger.

Ref T20690

Test Plan: Added sparse_embedding_tied_matmul with tests generated by cmake.

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Subscribers: godfrey.da.costa

Maniphest Tasks: T20690

Differential Revision: https://phabricator.sourcevertex.net/D32618