Addition of xPartition, yPartition to elementwise sparse meta info subGroupentry
Summary:
Add 2 new elements - intended to indicate which x and y partition the sub-group's information belowngs to.
The data is presently unused but the modification doesn't break anything else.  In other words the data content may
need revising.  This is to be used to identify the x or y partiton which the subGroup belongs to in support of the
sparse embedding update vertex

Ref T27439

Test Plan: CI

Reviewers: tomm, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: tomm, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T27439

Differential Revision: https://phabricator.sourcevertex.net/D33152