Support multiple ILDs for replicated collectives
Summary:
When compiling a graph for multiple ILDs, we can no longer assume that each ILD contains all replicas. This changes the semantic of IPU-link-domain collective operations: instead of forming a ring over the entire set of replicas, we instead need to form several rings (one per ILD) that will perform the operation in parallel.

NOTE: We can safely assume that all ILDs are identical and the number of replicas is evenly divisible by the number of ILDs.

The major difference is that the collective world size should now be the number of replicas per ILD, **not** the total number of replicas (as it is today). Additionally, the cross-replica copies used to describe the copies for each step in the collectives should cover all of the copies required by all of the parallel ILD rings.

Ref T26108.

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samuelh

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samuelh

Subscribers: samuelh, samk, danielb, alfredb, emmanuel, richard

Maniphest Tasks: T26108

Differential Revision: https://phabricator.sourcevertex.net/D31992