Serialise sparse embedding update to solve temporary memory and bucket spill problems
Summary:
Apply serialisation to solve both temporary memory problems and update spilled buckets correctly for sparse embedding update.

Each tile has a source buffer and destination buffer assigned to it.  Buffers contain the dense slices information, offsets(indices) relating to the slices and a column partition.  Each serial step the buffer contents are exchanged between tiles in a given partition in a circular manner and the sparse data on that tile updated (sparse data stays still).

The process will:
a) Prime buffers with the dense slices, offsets and column partition information
b) Run Update on the buffer content
c) Exchange to tiles in the partition of rows and z
d) Repeat until data returns to its initial tile

e) Exchange to tiles in the columns partition.
f) Repeat b) to d)
g) repeat e), f) based on the metadata overflow information

Vertices were updated to accept the column partition tensor that is exchanged in this scheme.  Once Slice is updated to cope with buffer overflows then the C++ code for the 2 vertices could be re-shared, and some of the tests rationalised.

Fixes T27439

Test Plan: Addition of tests to cover some bucket spills for update only.  In particular the existing tests (with no weighted area) would pass before having correctly dealt with column padding.  The current tests will fail unless that is done correctly, and will fail if the column partition data propogation is in the wrong direction.

Reviewers: tomm, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: tomm, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T27439

Differential Revision: https://phabricator.sourcevertex.net/D33292