Improve compile time for pooling operations
Summary:
When introspective mapping is used there is a lot
more work to be done to find tiles on which to
map work. The optimisations are mostly related
to speed that up.

For some pooling layers with large kernel (see T27742)
a speedup of around 25x is seen with this change on
some platforms.

Fixes T27799

Reviewers: tomm, timp, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: tomm, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T27799

Differential Revision: https://phabricator.sourcevertex.net/D33380