Serialising sparse embedding Slice to reduce temporary memory and Slice from spilled buckets
Summary:
Serialise the process (which was a 2-step process).

The general plan implemented is to exchange NZ and metaData from tile to tile, and run Slice vertices after every exchange step.  In this way the slice results are built up in memory step by step and never need to be exchanged themselves.  The sparse data "rows" can be split over sparse data X and Z partitions, and the sparse data is fully cycled step by step over the combination of these 2 partitions.
This deals with:
a) Row data being spread over the Z-partition - always the case if there are and Z-splits
b) Row data being spread over the X-partition - when there are bucket spills in this direction
c) The fact that the indices are mapped over multiple tiles in terms of their row allocations.  As each index is dynamic (and therefore unrelated to the sparse row data stored on its tile) this would require 2 stages in the dense multislice operation, however here the same problem is solved by serialised steps as described.

After each exchange over the full X,Z partition a column (or Y) partition exchange is performed and the whole process repeated.  This is to catch spilled data in that column, or Y dimension. The metadata indicated the number of times htis should be repeated.

** NOTE ** - After this some rationalisation was possible.  A few functions had parameters that were no longer needed (Needed for the 2-stage process that was applied) so these were removed.  The codelet code once more became common.  There are no real changes to the Update function other than those cause by the rationalisation.

Fixes T27438

Test Plan: Allow bucket spill for all tests that existed before, now for slice as well as update.  Merged test loops so that all tests have a likelihood of bucket spills

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Maniphest Tasks: T27438

Differential Revision: https://phabricator.sourcevertex.net/D33441