Sparse embedding slice assembler implementation for float and half
Summary:
Convert the slice vertex for sparse embedding to a  supervisor vertex and make an assembler implementation.

Similarly to the embedding matmul functions the supervisor processes buckets and the meta data found in them until a subGroup (in this case indicated by the yPartition) that the vertex needs to process is found.  Then the workers are used to process it.
Work division is done by assigning a sub-set of the offsets to each worker as that's what gets written to - so clashes are avoided.

As much pre-processing of vertex state as possible is done by the supervisor as possible.  This includes divide to divide work, but no individual calculation involving worker IDs.

It should be possible to save cycles by
Considering supervisor pipeline hits, and rearranging code to cope.  Also by placing the Sync after runall in a better place.  I'm intentionally leaving these for now as the code should be shared by the Update vertex and may be reworked.
After the first call we could enter the worker code after values are computed and put on the stack.  They are done by the worker as the calculation is worker specific but will be the same each time.
Some of the above things only make a difference if there is >1 subGroup in a bucket that actually gets processed

Yes, there are no cycle estimates!  This looks to be the case for many of the sparse mautmul vertices too.  Of course as the data content is dynamic and dictates the cycle cost an accurate estimate isn't possible.

Fixes T27435

Test Plan: More test variants added, intended to exercise the offsets work division in particular.

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm, oleksiik, samk

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Maniphest Tasks: T27435

Differential Revision: https://phabricator.sourcevertex.net/D33663