Decide the partial types for each level based on the convVertexType
Summary:
Previously we determined the partial types before knowing the
convVertexType. This was sub-optimial if the output type for vertex
was smaller than the partial type in the options which may happen if
the output of the vertex is the final result.

This change means we can no longer reuse the same types across the forward / backward /
weight update passes of a joint plan and must instead recalculate them
based on the vertex type for each pass.

Ref T27617

Reviewers: samk, oleksiik, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: oleksiik, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Subscribers: jamesn

Maniphest Tasks: T27617

Differential Revision: https://phabricator.sourcevertex.net/D33491