For reductions, detect overall column ordering and re-order input, output Tensors to avoid costly re-arrangement
Summary:
Sorry - a lot of JUST REARRANGEMENT.  The new file `ReductionIntrospection.cpp` contains new functions after and including `allTileGroupedPartials`.  Previous functions are those extracted from `ReductionStages.cpp`.
The same is true for much of `ReductionIntrospection.hpp`.  I made minor changes in those functions (Didn't intend to when I started) with emplace_back and types of loops.  Nothing functionally different though.
The function `allTileGroupedPartials` is also just a copy of what was in a loop.

EDIT - A separate speed improvement applied in those re-arranged functions, see this.  https://phabricator.sourcevertex.net/D30895?vs=101585&id=101851#toc . In which, the original vector[columns][Instance of column on tile] could be huge, eg with 20M columns but only 32k of them populated on a tile.  Storing structures only for the columns observed and then sorting resulted in a big speedup - pattern analysis was around 1 min 15 secs and became around 5 secs.

What's actually new then ?!
Moved pattern analysis to the top level, and as it is already done, pass that into the functions that need it in `ReductionStages.cpp`.  Examine that pattern analysis to see if the relative ordering of columns found in a region on a tile (The ordering that matters to the reduction library) is consistent for a given column over all tiles.  This detects the type of pattern that typically happens when dimensions are shuffled after a tensor is  created.  Re-order the input and output so that a simpler sequence of incremental columns is presented to the reduction stages, notably the intermediate partials representation.  Reduction then happens without the need for the rearrangement unless we are reducing withOutput in which case a rearrangement can happen at the end, although this works on fewer elements than it would have to rearrange otherwise.

Fixes T15113

Test Plan:
Specific tests added for the new column ordering analysis.
Tests added that cause this to be applied where there are various stages of reduction expected.
Verified to fix one real use example - T15085, the planner has apparently changed and we need to append `--plan-constraints={\"lookupSplit\":\"2\"}` to the command given.  This problem is fixed but that revealed a slow compile time, see above.
Verified over CI tests.

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm, samk, samuelh

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Maniphest Tasks: T15113

Differential Revision: https://phabricator.sourcevertex.net/D30895