Optimise sparse multi update add vertex in assembler
Summary:
A supervisor vertex, using the gradW meta information to determine work division between the workers.  Needs that to be generated in the meta info.
Made the vertex that processes half data use a float output that then gets cast and added to the actual NZ data.  This should help maintain resolution of the data during update where one NZ entry can get updated by many offsets.  Also made the scale variable a float regardless of the input type as there is no implementation benefit to it being half.

Fixes T27437

Test Plan: Further tests added, providing different baseT sizes as work is split based on the metainfo content and varying the size affects the content.  It is a bit ad-hoc though as control of the metainfo is not direct.

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm, oleksiik, samk

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Maniphest Tasks: T27437

Differential Revision: https://phabricator.sourcevertex.net/D33953