Add codelets for embedding slice, update with block sparse tensors, integrate into poplibs sparse embedding functions
Summary:
Added block sparse versions of the `SparseDenseMultiSlice` and `SparseDenseMultiUpdateAdd` codelets, with the test adapted to deal with both block and elementWise sparse data.
Integrated the codelets into the `embeddingSlice` and `embeddingUpdateAdd` functions.  This helped to confirm that the same method of adding xPartition and yPartition to the metaData  as is used for elementWise sparse data is the way to go (Already landed).  The codelets use that metadata. So the embedding layer is functional with C++ codelets and uses the same general method as the elementWise version.

In principle any block size can be supported when doing embedding slice and update, even non-square blocks and those with odd numbers of rows, columns.  However if the number of columns is a multiple of the vectorwidth of the data type, the copy or update-add operation can be vectorised. The intention is that a vectorised vertex could be implemented in assembler that would take advantage of this.

Ref T27823

Test Plan: Tests for both the codelets and the embedding layer added for Block sparsity.  Codelet tests to be revised to cover whatever parameters the assembler versions require

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Maniphest Tasks: T27823

Differential Revision: https://phabricator.sourcevertex.net/D34302