Assembler codelets for checking NaN in a tensor
Summary:
Use accumulator to process 8 elements in the innermost
loop for fp16 and 4 for fp32. Bank constraints are not
imposed and performance can thus be improved further
by a factor of 2.

Ref T28908

Test Plan:
Add new tests to use a slice of a tensor to check
for different remainders.

Reviewers: samk, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: samk, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T28908

Differential Revision: https://phabricator.sourcevertex.net/D34635