Move TANH, RELU, SIGMOID from popnn to popops
Summary:
The popnn source files NonLinearitySupervisor.S.in and NonLinearity2D.S.in
that define optimized, assembly in-place vertices for TANH, SIGMOID and
RELU, are moved to popops, renaming them as
unary2D/SupervisorNonLinearity.S.in.

A new popops::UnaryOpType::RELU operation is created.

New templates are created in elemwiseUnaryCodelets.cpp to accomodate the
different state fields of those vertices.

The creation of the vertices is appropriately modified in
popnn/NonLinearity.cpp

New test code is added.

Reviewers: tomm, timp, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: timp, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T13716

Differential Revision: https://phabricator.sourcevertex.net/D34409