Use DebugContext on api
Summary:
This diff replaces the `std::string` debug name on the poplibs api with the `DebugContext`. Internally the `DebugContext` is turned into a string to keep the internal code unchanged in this diff.

This has been done so the api is not a mix of approached.

Also the highlevel software can start integrating the DebugContext, in preparation of the internals changing.

Changing the poplibs internals to use DebugContext will be done in small incremental steps.

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, hectorm

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, hectorm

Maniphest Tasks: T29192

Differential Revision: https://phabricator.sourcevertex.net/D34860