Assembler implementation of sparse multislice vertices for block sparsity
Summary:
Implementation of half, float is quite similar but macros are used to create completely distinct vertices
with no code sharing (I don't expect multiple vertsions to exist in the same graph). Takes advantage of
the way that each row is always a multiple of 4 bytes. Support for actual data copy where the block has
data width that is multiple of 2, 4 or 8 bytes.  2bytes for half only. 4,8 for float and half.

Ref T27823

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Maniphest Tasks: T27823

Differential Revision: https://phabricator.sourcevertex.net/D34987