Deprecating use-unstable-format.
Summary:
Instead of use-unstable-format, the new way of setting the profile
format is with profile-format. The tools that were using the
unstable format are now using the experimental format (v2).
As the unstable format now points to v3 (SQLite), the tests have
to use v2 or be updated to adapt to v3.

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, Richardc, simonl

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, Richardc

Maniphest Tasks: T28814

Differential Revision: https://phabricator.sourcevertex.net/D35028