Fixed non-mapped tensor C when creating a graph
Summary:
Empty tensor C was created during the graph partitioning. It was not mapped to any tiles.
As a result if user created BsMatMulParams object and *did not* call bsMatMul() on it later, the program will give "not all tiles are mapped" error.
This is dangerous and can lead to difficult to discover bugs.

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shengf, samk, richard

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shengf, samk

Maniphest Tasks: T23552

Differential Revision: https://phabricator.sourcevertex.net/D34959