Adding option to control the position of gru reset gate
Summary:
Our current implementation of GRU is based on this paper: https://arxiv.org/abs/1406.1078v3 where the reset gate is applied before the candidate weights and biases.
There is an alternate implementation in this paper https://arxiv.org/abs/1406.1078v1 where the reset gate is appplied after the candidate weights and biases. This allows the weights for all three gates to be applied in a single matmul.

This revision adds the alternate implementation to the poplibs GRU. A resetAfter added to the interfaces controls which implementation is used, equivalent to the reset_after flag in keras.

Like in keras, when resetAfter=true, two sets of biases are used, one for the input components, and one for the recurrent components.

Test Plan: Added tests which run through all layers with resetAfter=true and compare results to a reference cpu implementation

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk, tomm, richard, godfrey.da.costa, Eddy

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Subscribers: georgep, samk

Maniphest Tasks: T25071

Differential Revision: https://phabricator.sourcevertex.net/D33731