Make loop variable referenceness consistent
Summary:
Clang has a new warning (which we set to `-Werror`) that makes you not use a reference for loop variables if it would result in a copy anyway, and also not use a copy if you could use a reference.

This is better than just using references everywhere because it makes it obvious where copies are happening, and could potentially be avoided.

Unfortunately it doesn't really work that well in templates, so I had to disable it in a couple of places.

Fixes T29445

Test Plan: Poplibs compiles on my Mac.

Reviewers: samk, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, mattl

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, mattl

Subscribers: mattl

Maniphest Tasks: T29445

Differential Revision: https://phabricator.sourcevertex.net/D35093