Implement MultiUpdateAdd vertices for block-sparse data in assembler
Summary:
Assembler versions of the vertices - Similar to the elementWise vertices work is divided using the metaInfo for gradW.
There are 2 versions of the vertex, one which copies using float/half vectors and another that uses scalar values.
This doesn't make a massive difference as searching through metaInfo can take most of the time, but seems worthwhile.  The
vector width is 2 always (even for half) as casting to float means there is no benefit in using half4

Fixes T27823

Reviewers: tomm, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: tomm, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T27823

Differential Revision: https://phabricator.sourcevertex.net/D34839