Consider AMP vertices with half partials if there is no reduction after
Summary:
If there is no reduction after the AMP vertex (i.e. the output is the
final result) then we can use a smaller type than the partial type
requested by the user. This may eliminate a unnecessary cast or, on Mk2
allow us to use double AMP.

There was already logic along these lines for the backward and weight
update passes of joint plans, but only if
getFp16InFp16OutConvUnitsPerTile() == getFp16InFp32OutConvUnitsPerTile()
which meant it didn't happen on Mk2. With this change we apply the same
logic to non-joint plans and to both Mk1 and Mk2.

getConvVertexTypeCandidates() has been updated to return AMP candidates
with smaller partial types if it doesn't affect the result (it already
did this for outer product). For joint plans we call
getConvVertexTypeCandidates() to get the list of candidates for the weight
update and backward passes and use the one at the beginning of the list.
If we can use a smaller partial type that candidate will be at the head of
the list (since candidates are ordered from most to least promising).

Fixes T27617

Depends on D34057

Reviewers: samk, oleksiik, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: samk, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T27617

Differential Revision: https://phabricator.sourcevertex.net/D34096