Limit test values range to avoid HALF overflow
Summary:
T29575 is caused by the fact that the values that are used during these tests
might generate an HALF overflow.
So we limit the HALF values of the two operands of a binary operation between
-200 and +200 (this was -1000..+1000, that would cause a MULTIPLY to overflow).
For DIVIDE we set the dividend in the range -200 and +200 and the divisor in
the range 0.5..600, so that the result absolute value is between 0.0 and 400.0.
Note that POWER was already limited.

Test Plan: poplibs regression test

Reviewers: oleksiik, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: oleksiik, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Subscribers: samk

Maniphest Tasks: T29575

Differential Revision: https://phabricator.sourcevertex.net/D35388