Use memoise function to implement the cache for the pool plan cycle estimates
Summary:
As suggested (Ref T29219) this will use the memoise function to cache cycle estimate results for the pooling planner.

It functions but I've not yet measured a speed increase.  Thinking back, the cacheing of results probably wasn't the real cost of slow compilation.  So, although this change is complete more investigation could show that there are other gains to be made.

Test Plan: CI

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Maniphest Tasks: T29219

Differential Revision: https://phabricator.sourcevertex.net/D35504