Change tolerance used in implicit conversion from half to float in scaledAdd to 0.1%
Summary:
The default tolerance used in scaled add to implicitly
convert half operations to float uses tolerance way
below the error in half type. This would mean that on
the average we could end up choosing float operations
most of the time. Use a tolerance that is much closer
to the a value slightly greater than rounding errors
in half.

Shows signficant speed up in norms for EfficientNet for
example with some layer gradients taking 10% less cycles.

Ref T29772

Reviewers: timp, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: timp, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Subscribers: samk

Maniphest Tasks: T29772

Differential Revision: https://phabricator.sourcevertex.net/D35602