Generalise assembler for binary and broadcast ops and implement short, unsigned short OR and AND as 4x16 bit vectorised loops
Summary:
Made the macro system for binary and broadcast ops more generic:
The processing for the "half" loop for add/mul/sub is very similar to that needed to process short, unsigned short OR and AND instructions (There is an ARF OR64 and AND64 instruction).  So the loop and 1of, 2of processing stubs were made more generic with macros accepting a type (half, short, unsigned_short) and the loop structure having a more generic name (4x16bit instead of half).
The instruction is passed into the macro, instead of the macro adding the `f16v4\OPERATION` syntax.

Once made more flexible the same macros can be used for short and unsigned short as for half.
Apart from a few minor naming changes and passing the instructions in for consistency float operation is left the same.
Added a small optimisation to decrement using brnzdec instead of subtract.

One complication is that there is no ARF OR32 or AND32 instruction. This means continuing to use OR64 and AND64 where the over processing is OK for bitwise operations as they don't cause exceptions - however the register assignment for function stub output had to be revised as a result

Fixes T29267

Test Plan: CI

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, claudiop, oleksiik

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oleksiik

Maniphest Tasks: T29267

Differential Revision: https://phabricator.sourcevertex.net/D35629