Add tests for grad only non-linearity (as non-linearity is a unary op). Modify assembler to use macros not cmake
Summary:
The 3 non linearities : relu, tanh and sigmoid were recently moved to be unary ops, and the test for them removed from popnn.  That left the grad variants of the codelets untested - so this diff makes that possible by allowing the test to check just grad or fwd (Which is needed as the fwd codelets don't exist anymore in popnn).

With proper testing, converted the assembler file that implemented these codelets to use macros instead of cmake as the way to create 3 codelets from one set of source code.  This is in line with most of the codebase and should be more readable

Fixes T29948

Test Plan: relu, tanh and sigmoid 2D and supervisor tests added to cmake

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, claudiop, oleksiik, tomm

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, claudiop

Maniphest Tasks: T29948

Differential Revision: https://phabricator.sourcevertex.net/D35858